finalized on google sign in and share auth with android things kit using nearby connection

+++ app/src/main/java/ke/co/appslab/smarthome/models/Weather.kt
@@ -1,6 +1,6 @@
 package ke.co.appslab.smarthome.models
 
 data class Weather(
-    val temperature: Float = 0f,
-    val pressure: Float = 0f
+    val temperature: Int = 0,
+    val pressure: Int = 0
 )
